requirements.txt
streamlit
opencv-python
numpy
pandas
seaborn

albumentations
pytesseract
scikit-image
pillow
timm